+++ addons/misc/script_component.hpp
@@ -15,3 +15,4 @@
 #endif
 
 #include "\x\kat\addons\main\script_macros.hpp"
+#include "\z\ace\addons\medical_engine\script_macros_medical.hpp"
