+++ addons/pharma/functions/fnc_checkIV.sqf
@@ -20,8 +20,9 @@
 params ["_medic", "_patient", "_bodyPart"];
 
 if !(GVAR(ivCheckLimbDamage)) exitWith {
-    true;
+    true
 };
+
 private _bodyPartDamage = _patient getVariable ["ace_medical_bodyPartDamage", [0, 0, 0, 0, 0, 0]];
 private _compare = ALL_BODY_PARTS find toLower _bodyPart;
 private _final2 = 0;
