+++ addons/pharma/functions/fnc_treatmentAdvanced_ReorientationLocal.sqf
@@ -18,7 +18,7 @@
 
 params ["_medic", "_patient"];
 
-private _sedated = _patient getVariable [QGVAR(sedated), false];
+private _sedated = _patient getVariable [QEGVAR(surgery,sedated), false];
 if (_sedated) exitWith {};
 
 private _bloodVolume = GET_BLOOD_VOLUME(_patient);
